+++ lib/widget/loader.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class Loader extends StatelessWidget {
+  const Loader({
+    Key key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(
+        child: CircularProgressIndicator(
+          backgroundColor: Theme.of(context).primaryColor,
+        ),
+      ),
+    );
+  }
+}
