+++ lib/widget/widgets.dart
@@ -6,3 +6,4 @@ export 'drawerButton.dart';
 export 'dateTimePicker.dart';
 export 'placeholder_widget.dart';
 export 'smallButton.dart';
+export 'loader.dart';
